+++ src/main/java/org/fuin/utils4j/package-info.java
@@ -19,4 +19,7 @@
 /**
  * Basic Java utilities.
  */
+@NullMarked
 package org.fuin.utils4j;
+
+import org.jspecify.annotations.NullMarked;