+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes
 
+## 0.16.0
+- Added new [Utils4J](src/main/java/org/fuin/utils4j/Utils4J.java) method to encode a UUID to a short string and back
+- Added [JSpecify](https://jspecify.dev/) and [NullAway](https://github.com/uber/nullaway)
+
 ## 0.15.0
 - Added new [JandexUtils](src/main/java/org/fuin/utils4j/jandex/JandexUtils.java) methods
 - Added new [TechnicalId](src/main/java/org/fuin/utils4j/TechnicalId.java) tag interface.